Extract createSchemaFromLcoation function from serve command
